requirements.txt
PyQt5==5.15.11
anthropic>=0.88.0
python-dotenv>=1.2.2
robosuite>=1.5.2
numpy>=1.26.4
